requirements.txt
PyPDF2
sentence_transformers
pinecone-client[grpc]
datasets